Conflicts:
	APPWEB-INCOMEMotorCalculoJava-api/pom.xml
	APPWEB-INCOMEMotorCalculoJava-boot/pom.xml
	APPWEB-INCOMEMotorCalculoJava-boot/src/main/resources/application-standalone.yml
	APPWEB-INCOMEMotorCalculoJava-boot/src/main/resources/query/primary/common.xml
	APPWEB-INCOMEMotorCalculoJava-model/pom.xml
	APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/app/run/tarea/procesar/async/service/RunTareaProcesarVentaAsyncServiceImpl.java
	APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/app/run/tarea/procesar/service/RunTareaProcesarVentaServiceImpl.java
	APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaLocalizacionVentaRepositoryCustom.java
	APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/tarea/repository/TareaLocalizacionVentaRepositoryCustomImpl.java
	APPWEB-INCOMEMotorCalculoJava-ms/pom.xml
	APPWEB-INCOMEMotorCalculoJava-ws/pom.xml
	pom.xml

+++ APPWEB-INCOMEMotorCalculoJava-api/src/main/java/com/inditex/rrhh/icmclcwb/api/app/tarea/dto/TareaTipoHoraDto.java
@@ -17,7 +17,7 @@ public class TareaTipoHoraDto implements Serializable {
 	
 	@NotBlank
 	@ApiModelProperty(value = "Id de tipo de hora", required = true)
-	private Long idTipoHora;
+	private Integer idTipoHora;
 
 	@NotNull
 	@ApiModelProperty(value = "Flag excluido calculo", required = true)
